Compacted the XML loading code and move it into a separate package.
