+++ resources/fxml/LessonPane.fxml
@@ -8,7 +8,7 @@
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
 
-<TitledPane fx:id="namePane" text="Name" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.meti.LessonPane">
+<TitledPane fx:id="namePane" text="Name" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.meti.lesson.LessonPane">
    <content>
       <BorderPane prefHeight="200.0" prefWidth="200.0">
          <center>
